Move OutputResolver class
